Permite apenas admins acessarem a tela de usuários

+++ src/app/Http/Controllers/UsuarioController.php
@@ -10,13 +10,19 @@ use App\Models\Usuario\Usuario;
 use Exception;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
+
+use function Laravel\Prompts\error;
 
 class UsuarioController extends Controller
 {
 
-    public function index()
+    public function index(Request $request)
     {
+        $this->authorize('usuarios', Usuario::class);
+
         $usuarios = Usuario::with('user')->paginate(10);
 
         return view('usuario.index', ['usuarios' => $usuarios]);
@@ -24,6 +30,7 @@ class UsuarioController extends Controller
 
     public function create()
     {
+        $this->authorize('usuarios', Usuario::class);
         $tipos = TipoUsuario::getAll();
 
         return view('usuario.create', ['tipos' => $tipos]);
@@ -31,6 +38,7 @@ class UsuarioController extends Controller
 
     public function salvar(UsuarioPostRequest $request)
     {
+        $this->authorize('usuarios', Usuario::class);
         $campos = $request->validated();
         $usuario = Usuario::create($campos);
 
@@ -51,7 +59,7 @@ class UsuarioController extends Controller
 
     public function view(int $id)
     {
-
+        $this->authorize('usuarios', Usuario::class);
         $usuario = $this->findUsuario($id);
 
         return view('usuario.view', ['usuario' => $usuario]);
@@ -59,6 +67,7 @@ class UsuarioController extends Controller
 
     public function update(Usuario $usuario)
     {
+        $this->authorize('usuarios', Usuario::class);
         $tipos = TipoUsuario::getAll();
 
         return view('usuario.update', ['usuario' => $usuario, 'tipos' => $tipos]);
@@ -66,7 +75,7 @@ class UsuarioController extends Controller
 
     public function atualizar(UsuarioPostUpdateRequest $request, Usuario $usuario)
     {
-
+        $this->authorize('usuarios', Usuario::class);
         // dd($request->validated());
         $usuario->update($request->validated());
 
@@ -82,10 +91,11 @@ class UsuarioController extends Controller
 
     public function destroy(Usuario $usuario)
     {
-        try{
+        $this->authorize('usuarios', Usuario::class);
+        try {
             $usuario->delete();
-        } catch(Exception $e){
-            return redirect()->route('/usuario')->with('danger','Não foi possível excluir este usuário.');
+        } catch (Exception $e) {
+            return redirect()->route('/usuario')->with('danger', 'Não foi possível excluir este usuário.');
         }
         return redirect()->route('/usuario');
     }
@@ -93,6 +103,7 @@ class UsuarioController extends Controller
 
     private function findUsuario(int $id)
     {
+        $this->authorize('usuarios', Usuario::class);
         try {
             $usuario = Usuario::find($id);
             if ($usuario) {
